Aggiornamento: il file di testo viene letto e per ogni nota letta viene creato un oggetto Note che ne raccoglie le caratteristiche. L'ideale sarebbe formare una lista di Note per poi iterarci agevolmente.

+++ untitled1/src/Note.scala
@@ -1,4 +1,4 @@
-class Note (var pitch: Float, var time: Float) {
+class Note (var pitch: Int, var time: Float) {
   override def toString: String =
     s"(Altezza MIDI: $pitch, Durata: $time)"
 }
